Struct admin1_directory::Directory

pub struct Directory<C, NC, A> {
    // some fields omitted
}

Central instance to access all Directory related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate admin1_directory;
use admin1_directory::Channel;
use admin1_directory::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use admin1_directory::Directory;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Channel = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().watch(&req)
             .view_type("et")
             .sort_order("diam")
             .show_deleted("ipsum")
             .query("Lorem")
             .projection("et")
             .page_token("duo")
             .order_by("aliquyam")
             .max_results(-9)
             .event("Lorem")
             .domain("eos")
             .customer("erat")
             .custom_field_mask("sadipscing")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Methods

impl<'a, C, NC, A> Directory<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn new(client: C, authenticator: A) -> Directory<C, NC, A>

fn asps(&'a self) -> AspMethodsBuilder<'a, C, NC, A>

fn channels(&'a self) -> ChannelMethodsBuilder<'a, C, NC, A>

fn chromeosdevices(&'a self) -> ChromeosdeviceMethodsBuilder<'a, C, NC, A>

fn groups(&'a self) -> GroupMethodsBuilder<'a, C, NC, A>

fn members(&'a self) -> MemberMethodsBuilder<'a, C, NC, A>

fn mobiledevices(&'a self) -> MobiledeviceMethodsBuilder<'a, C, NC, A>

fn notifications(&'a self) -> NotificationMethodsBuilder<'a, C, NC, A>

fn orgunits(&'a self) -> OrgunitMethodsBuilder<'a, C, NC, A>

fn schemas(&'a self) -> SchemaMethodsBuilder<'a, C, NC, A>

fn tokens(&'a self) -> TokenMethodsBuilder<'a, C, NC, A>

fn users(&'a self) -> UserMethodsBuilder<'a, C, NC, A>

fn verification_codes(&'a self) -> VerificationCodeMethodsBuilder<'a, C, NC, A>