Struct admin1_directory::UserMethodsBuilder

pub struct UserMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the Directory hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate admin1_directory;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Directory::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `aliases_delete(...)`, `aliases_insert(...)`, `aliases_list(...)`, `aliases_watch(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `make_admin(...)`, `patch(...)`, `photos_delete(...)`, `photos_get(...)`, `photos_patch(...)`, `photos_update(...)`, `undelete(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, NC, A> UserMethodsBuilder<'a, C, NC, A>

fn insert(&self, request: &User) -> UserInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

create user.

fn photos_delete(&self, user_key: &str) -> UserPhotoDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Remove photos for the user

fn undelete(&self, request: &UserUndelete, user_key: &str) -> UserUndeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Undelete a deleted user

fn make_admin(&self, request: &UserMakeAdmin, user_key: &str) -> UserMakeAdminMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

change admin status of a user

fn patch(&self, request: &User, user_key: &str) -> UserPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

update user. This method supports patch semantics.

fn watch(&self, request: &Channel) -> UserWatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Watch for changes in users list

fn delete(&self, user_key: &str) -> UserDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete user

fn photos_update(&self, request: &UserPhoto, user_key: &str) -> UserPhotoUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Add a photo for the user

fn get(&self, user_key: &str) -> UserGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

retrieve user

fn aliases_insert(&self, request: &Alias, user_key: &str) -> UserAliaseInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Add a alias for the user

fn list(&self) -> UserListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieve either deleted users or all users in a domain (paginated)

fn aliases_delete(&self, user_key: &str, alias: &str) -> UserAliaseDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Remove a alias for the user

fn update(&self, request: &User, user_key: &str) -> UserUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

update user

fn aliases_watch(&self, request: &Channel, user_key: &str) -> UserAliaseWatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Watch for changes in user aliases list

fn aliases_list(&self, user_key: &str) -> UserAliaseListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List all aliases for a user

fn photos_get(&self, user_key: &str) -> UserPhotoGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieve photo of a user

fn photos_patch(&self, request: &UserPhoto, user_key: &str) -> UserPhotoPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Add a photo for the user. This method supports patch semantics.