Struct admin1_directory::UserOrganization

pub struct UserOrganization {
    pub domain: Option<String>,
    pub description: Option<String>,
    pub title: Option<String>,
    pub symbol: Option<String>,
    pub custom_type: Option<String>,
    pub primary: Option<bool>,
    pub location: Option<String>,
    pub cost_center: Option<String>,
    pub department: Option<String>,
    pub type_: Option<String>,
    pub name: Option<String>,
}

JSON template for an organization entry.

This type is not used in any activity, and only used as part of another schema.

Fields

domain

The domain to which the organization belongs to.

description

Description of the organization.

title

Title (designation) of the user in the organization.

symbol

Symbol of the organization.

custom_type

Custom type.

primary

If it user's primary organization.

location

Location of the organization. This need not be fully qualified address.

cost_center

The cost center of the users department.

department

Department within the organization.

type_

Each entry can have a type which indicates standard types of that entry. For example organization could be of school, work etc. In addition to the standard type, an entry can have a custom type and can give it any name. Such types should have the CUSTOM value as type and also have a CustomType value.

name

Name of the organization